backend/requirements.txt
openai-whisper>=20230314
pyaudio>=0.2.13
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

